refactor: split oversized files by responsibility
