+++ internal/bridge/bridge.go
@@ -2,102 +2,15 @@ package bridge
 
 import (
 	"context"
-	"crypto/rand"
-	"encoding/json"
 	"fmt"
-	"log/slog"
-	"math/big"
-	"os"
 	"sync"
 	"time"
 
-	"github.com/chromedp/cdproto/page"
 	"github.com/chromedp/cdproto/target"
-	"github.com/chromedp/chromedp"
 	"github.com/pinchtab/pinchtab/internal/config"
 	"github.com/pinchtab/pinchtab/internal/ids"
 	"github.com/pinchtab/pinchtab/internal/stealth"
 )
-
-type TabEntry struct {
-	Ctx                   context.Context
-	Cancel                context.CancelFunc
-	Accessed              bool
-	CDPID                 string    // raw CDP target ID
-	CreatedAt             time.Time // when the tab was first created/registered
-	LastUsed              time.Time // last time the tab was accessed via TabContext
-	Policy                TabPolicyState
-	Watching              bool
-	ConsoleCaptureEnabled bool
-
-	// Lifecycle auto-close timer. autoCloseGen is bumped on every (re)schedule
-	// so a fire that races with a reset/cancel can detect itself and bail.
-	autoCloseTimer *time.Timer
-	autoCloseGen   uint64
-}
-
-type RefTarget struct {
-	BackendNodeID  int64  `json:"backendNodeId"`
-	FrameID        string `json:"frameId,omitempty"`
-	FrameURL       string `json:"frameUrl,omitempty"`
-	FrameName      string `json:"frameName,omitempty"`
-	ChildFrameID   string `json:"childFrameId,omitempty"`
-	ChildFrameURL  string `json:"childFrameUrl,omitempty"`
-	ChildFrameName string `json:"childFrameName,omitempty"`
-}
-
-type RefCache struct {
-	Refs    map[string]int64
-	Targets map[string]RefTarget
-	Nodes   []A11yNode
-}
-
-func (c *RefCache) Lookup(ref string) (RefTarget, bool) {
-	if c == nil {
-		return RefTarget{}, false
-	}
-	if c.Targets != nil {
-		if target, ok := c.Targets[ref]; ok {
-			return target, true
-		}
-	}
-	if c.Refs != nil {
-		if nid, ok := c.Refs[ref]; ok {
-			return RefTarget{BackendNodeID: nid}, true
-		}
-	}
-	return RefTarget{}, false
-}
-
-func RefTargetsFromNodes(nodes []A11yNode) map[string]RefTarget {
-	targets := make(map[string]RefTarget, len(nodes))
-	for _, node := range nodes {
-		if node.Ref == "" || node.NodeID == 0 {
-			continue
-		}
-		targets[node.Ref] = RefTarget{
-			BackendNodeID:  node.NodeID,
-			FrameID:        node.FrameID,
-			FrameURL:       node.FrameURL,
-			FrameName:      node.FrameName,
-			ChildFrameID:   node.ChildFrameID,
-			ChildFrameURL:  node.ChildFrameURL,
-			ChildFrameName: node.ChildFrameName,
-		}
-	}
-	return targets
-}
-
-type FrameScope struct {
-	FrameID   string `json:"frameId,omitempty"`
-	FrameURL  string `json:"frameUrl,omitempty"`
-	FrameName string `json:"frameName,omitempty"`
-	OwnerRef  string `json:"ownerRef,omitempty"`
-}
-
-func (s FrameScope) Active() bool {
-	return s.FrameID != ""
-}
 
 type Bridge struct {
 	AllocCtx      context.Context
@@ -159,7 +72,6 @@ func New(allocCtx, browserCtx context.Context, cfg *config.RuntimeConfig) *Bridg
 	}
 	b.ensureStealthBundle()
 	b.Dialogs = NewDialogManager()
-	// Only initialize TabManager if browserCtx is provided (not lazy-init case)
 	if cfg != nil && browserCtx != nil {
 		b.TabManager = NewTabManager(browserCtx, cfg, idMgr, logStore, b.tabSetup)
 		b.SetOnAfterClose(func() { go b.SaveState() })
@@ -172,68 +84,6 @@ func New(allocCtx, browserCtx context.Context, cfg *config.RuntimeConfig) *Bridg
 	b.Locks = NewLockManager()
 	b.InitActionRegistry()
 	return b
-}
-
-func (b *Bridge) quietStealthObservers() bool {
-	return b != nil && b.Config != nil && stealth.NormalizeLevel(b.Config.StealthLevel) == stealth.LevelFull
-}
-
-func (b *Bridge) RestartStatus() (bool, time.Duration) {
-	if b == nil {
-		return false, 0
-	}
-	b.initMu.Lock()
-	defer b.initMu.Unlock()
-	if !b.draining {
-		return false, 0
-	}
-	remaining := time.Until(b.drainUntil)
-	if remaining < 0 {
-		remaining = 0
-	}
-	return true, remaining
-}
-
-func (b *Bridge) injectStealth(ctx context.Context) {
-	if b.StealthBundle == nil || b.StealthBundle.Script == "" {
-		return
-	}
-	if err := chromedp.Run(ctx,
-		chromedp.ActionFunc(func(ctx context.Context) error {
-			_, err := page.AddScriptToEvaluateOnNewDocument(b.StealthBundle.Script).Do(ctx)
-			return err
-		}),
-	); err != nil {
-		slog.Warn("stealth injection failed", "err", err)
-	}
-}
-
-func (b *Bridge) applyTargetStealth(ctx context.Context) {
-	if b == nil || b.Config == nil {
-		return
-	}
-
-	ua := ""
-	if b.StealthBundle != nil {
-		ua = b.StealthBundle.LaunchUserAgent()
-	}
-
-	if err := chromedp.Run(ctx, chromedp.ActionFunc(func(ctx context.Context) error {
-		return stealth.ApplyTargetEmulation(ctx, b.Config, ua)
-	})); err != nil {
-		slog.Warn("stealth target emulation failed", "err", err)
-	}
-}
-
-func (b *Bridge) tabSetup(ctx context.Context) {
-	b.applyTargetStealth(ctx)
-	b.installWorkerStealthParity(ctx)
-	b.injectStealth(ctx)
-	if b.Config.NoAnimations {
-		if err := b.InjectNoAnimations(ctx); err != nil {
-			slog.Warn("no-animations injection failed", "err", err)
-		}
-	}
 }
 
 // StartNetworkCapture enables network monitoring for a specific tab.
@@ -361,292 +211,6 @@ func (b *Bridge) ClearErrorLogs(tabID string) {
 	}
 }
 
-func (b *Bridge) EnsureChrome(cfg *config.RuntimeConfig) error {
-	b.initMu.Lock()
-	defer b.initMu.Unlock()
-
-	if b.draining {
-		return ErrBrowserDraining
-	}
-
-	if b.initialized && b.BrowserCtx != nil {
-		// Check if browser context is still alive
-		if b.BrowserCtx.Err() == nil {
-			return nil
-		}
-		// Chrome died — reset state for re-initialization
-		slog.Warn("chrome context cancelled, re-initializing")
-		b.initialized = false
-		b.BrowserCtx = nil
-		b.BrowserCancel = nil
-		b.AllocCtx = nil
-		b.AllocCancel = nil
-		b.TabManager = nil
-	}
-
-	if b.BrowserCtx != nil {
-		if b.BrowserCtx.Err() == nil {
-			return nil
-		}
-		b.BrowserCtx = nil
-		b.BrowserCancel = nil
-	}
-
-	slog.Debug("ensure chrome called", "headless", cfg.Headless, "profile", cfg.ProfileDir)
-
-	// Initialize Chrome if not already done
-	if err := AcquireProfileLock(cfg.ProfileDir); err != nil {
-		if cfg.Headless {
-			// If we are in headless mode, we are more flexible.
-			// Instead of failing, we can use a unique temporary profile dir.
-			uniqueDir, tmpErr := os.MkdirTemp("", "pinchtab-profile-*")
-			if tmpErr == nil {
-				slog.Warn("profile in use; using unique temporary profile for headless instance",
-					"requested", cfg.ProfileDir, "using", uniqueDir, "reason", err.Error())
-				cfg.ProfileDir = uniqueDir
-				b.tempProfileDir = uniqueDir
-				// Re-acquire lock for the new temp dir (should always succeed)
-				_ = AcquireProfileLock(cfg.ProfileDir)
-			} else {
-				slog.Error("cannot acquire profile lock and failed to create temp dir", "profile", cfg.ProfileDir, "err", err.Error(), "tmpErr", tmpErr.Error())
-				return fmt.Errorf("profile lock: %w (temp dir failed: %v)", err, tmpErr)
-			}
-		} else {
-			slog.Error("cannot acquire profile lock; another pinchtab may be active", "profile", cfg.ProfileDir, "err", err.Error())
-			return fmt.Errorf("profile lock: %w", err)
-		}
-	}
-
-	slog.Info("starting chrome with confirmed profile", "headless", cfg.Headless, "profile", cfg.ProfileDir)
-	b.ensureStealthBundle()
-	allocCtx, allocCancel, browserCtx, browserCancel, launchMode, err := InitChrome(cfg, b.StealthBundle)
-	if err != nil {
-		return fmt.Errorf("failed to initialize chrome: %w", err)
-	}
-
-	b.AllocCtx = allocCtx
-	b.AllocCancel = allocCancel
-	b.BrowserCtx = browserCtx
-	b.BrowserCancel = browserCancel
-	b.initialized = true
-	b.stealthLaunchMode = launchMode
-
-	// Initialize TabManager now that browser is ready
-	if b.Config != nil && b.TabManager == nil {
-		if b.IdMgr == nil {
-			b.IdMgr = ids.NewManager()
-		}
-		if b.LogStore == nil {
-			b.LogStore = NewConsoleLogStore(1000)
-		}
-		b.TabManager = NewTabManager(browserCtx, b.Config, b.IdMgr, b.LogStore, b.tabSetup)
-		b.SetOnAfterClose(func() { go b.SaveState() })
-		b.SetDialogManager(b.Dialogs)
-		b.SetNetworkMonitor(b.netMonitor)
-		if !b.quietStealthObservers() {
-			b.StartBrowserGuards()
-		}
-	}
-
-	// Ensure action registry is populated (idempotent)
-	if b.Actions == nil {
-		b.InitActionRegistry()
-	}
-
-	// Restore tabs from previous session (if any saved state exists). Opt-in
-	// via instanceDefaults.tabPolicy.restore=true; otherwise tabs closed
-	// before shutdown stay closed.
-	if b.tempProfileDir == "" && b.Config != nil && b.Config.TabRestore {
-		b.RestoreState()
-	}
-
-	// Start crash monitoring
-	if !b.quietStealthObservers() {
-		b.MonitorCrashes(nil)
-	}
-
-	return nil
-}
-
-// Cleanup releases browser resources and removes temporary profile directories.
-// Must be called on shutdown to prevent Chrome process and disk leaks.
-func (b *Bridge) RestartBrowser(cfg *config.RuntimeConfig) error {
-	if cfg == nil {
-		cfg = b.Config
-	}
-	if cfg == nil {
-		return fmt.Errorf("runtime config is required")
-	}
-
-	const drainWindow = 2 * time.Second
-
-	b.initMu.Lock()
-	b.draining = true
-	b.drainUntil = time.Now().Add(drainWindow)
-	b.initMu.Unlock()
-
-	slog.Info("browser soft restart: draining requests before restart", "drain_window", drainWindow)
-	time.Sleep(drainWindow)
-
-	b.initMu.Lock()
-
-	if b.BrowserCancel != nil {
-		b.BrowserCancel()
-		slog.Info("browser soft restart: cancelled browser context")
-	}
-	if b.AllocCancel != nil {
-		b.AllocCancel()
-		slog.Info("browser soft restart: cancelled allocator context")
-	}
-
-	profileDir := ""
-	if b.tempProfileDir != "" {
-		profileDir = b.tempProfileDir
-	} else {
-		profileDir = cfg.ProfileDir
-	}
-	if profileDir != "" {
-		time.Sleep(200 * time.Millisecond)
-		killed := killChromeByProfileDir(profileDir)
-		if killed > 0 {
-			slog.Info("browser soft restart: killed surviving chrome processes", "count", killed, "profileDir", profileDir)
-		}
-		ClearChromeSessions(profileDir)
-	}
-	b.ClearSavedState()
-
-	if b.tempProfileDir != "" {
-		if err := os.RemoveAll(b.tempProfileDir); err != nil {
-			slog.Warn("failed to remove temp profile dir during restart", "path", b.tempProfileDir, "err", err)
-		} else {
-			slog.Info("removed temp profile dir during restart", "path", b.tempProfileDir)
-		}
-		b.tempProfileDir = ""
-	}
-
-	b.initialized = false
-	b.BrowserCtx = nil
-	b.BrowserCancel = nil
-	b.AllocCtx = nil
-	b.AllocCancel = nil
-	b.TabManager = nil
-	b.stealthLaunchMode = stealth.LaunchModeUninitialized
-
-	b.LogStore = NewConsoleLogStore(1000)
-	b.netMonitor = NewNetworkMonitor(DefaultNetworkBufferSize)
-	if cfg.NetworkBufferSize > 0 {
-		b.netMonitor = NewNetworkMonitor(cfg.NetworkBufferSize)
-	}
-	b.fingerprintMu.Lock()
-	b.fingerprintOverlays = make(map[string]bool)
-	b.fingerprintMu.Unlock()
-	b.workerStealthTargets = sync.Map{}
-	b.Dialogs = NewDialogManager()
-	b.Locks = NewLockManager()
-	b.Config = cfg
-
-	b.StealthBundle = nil
-	b.Actions = nil
-	b.InitActionRegistry()
-
-	b.draining = false
-	b.drainUntil = time.Time{}
-	b.initMu.Unlock()
-
-	if err := b.EnsureChrome(cfg); err != nil {
-		return err
-	}
-	b.CleanupSavedStateBackup()
-	return nil
-}
-
-func (b *Bridge) Cleanup() {
-	// Persist open tabs so next startup can restore them
-	if b.TabManager != nil && b.tempProfileDir == "" {
-		b.SaveState()
-	}
-
-	// Mark a clean exit so Chrome doesn't show a crash recovery bar
-	if b.Config != nil && b.tempProfileDir == "" {
-		MarkCleanExit(b.Config.ProfileDir)
-	}
-
-	// Cancel chromedp contexts (kills main Chrome process)
-	if b.BrowserCancel != nil {
-		b.BrowserCancel()
-		slog.Debug("chrome browser context cancelled")
-	}
-	if b.AllocCancel != nil {
-		b.AllocCancel()
-		slog.Debug("chrome allocator context cancelled")
-	}
-
-	// Chrome spawns helpers (GPU, renderer) in their own process groups.
-	// Context cancellation only kills the main process. Kill survivors
-	// by scanning for processes using our profile directory.
-	profileDir := ""
-	if b.tempProfileDir != "" {
-		profileDir = b.tempProfileDir
-	} else if b.Config != nil {
-		profileDir = b.Config.ProfileDir
-	}
-	if profileDir != "" {
-		// Brief wait for context cancel to propagate
-		time.Sleep(200 * time.Millisecond)
-		killed := killChromeByProfileDir(profileDir)
-		if killed > 0 {
-			slog.Info("cleanup: killed surviving chrome processes", "count", killed, "profileDir", profileDir)
-		}
-	}
-
-	if b.tempProfileDir != "" {
-		if err := os.RemoveAll(b.tempProfileDir); err != nil {
-			slog.Warn("failed to remove temp profile dir", "path", b.tempProfileDir, "err", err)
-		} else {
-			slog.Info("removed temp profile dir", "path", b.tempProfileDir)
-		}
-		b.tempProfileDir = ""
-	}
-}
-
-func (b *Bridge) SetBrowserContexts(allocCtx context.Context, allocCancel context.CancelFunc, browserCtx context.Context, browserCancel context.CancelFunc) {
-	b.initMu.Lock()
-	defer b.initMu.Unlock()
-
-	b.AllocCtx = allocCtx
-	b.AllocCancel = allocCancel
-	b.BrowserCtx = browserCtx
-	b.BrowserCancel = browserCancel
-	b.initialized = true
-	b.stealthLaunchMode = stealth.LaunchModeAttached
-
-	// Now initialize TabManager with the browser context
-	if b.Config != nil && b.TabManager == nil {
-		if b.IdMgr == nil {
-			b.IdMgr = ids.NewManager()
-		}
-		if b.LogStore == nil {
-			b.LogStore = NewConsoleLogStore(1000)
-		}
-		b.TabManager = NewTabManager(browserCtx, b.Config, b.IdMgr, b.LogStore, b.tabSetup)
-		b.SetOnAfterClose(func() { go b.SaveState() })
-		b.SetDialogManager(b.Dialogs)
-		b.SetNetworkMonitor(b.netMonitor)
-	}
-}
-
-func (b *Bridge) ensureStealthBundle() {
-	if b.StealthBundle != nil || b.Config == nil {
-		return
-	}
-	b.StealthBundle = stealth.NewBundle(b.Config, cryptoRandSeed())
-}
-
-func (b *Bridge) StealthStatus() *stealth.Status {
-	b.ensureStealthBundle()
-	return stealth.StatusFromBundle(b.StealthBundle, b.Config, b.stealthLaunchMode)
-}
-
 func (b *Bridge) SetFingerprintRotateActive(tabID string, active bool) {
 	if tabID == "" {
 		return
@@ -669,44 +233,6 @@ func (b *Bridge) BrowserContext() context.Context {
 	return b.BrowserCtx
 }
 
-func (b *Bridge) ExecuteAction(ctx context.Context, kind string, req ActionRequest) (map[string]any, error) {
-	kind = CanonicalActionKind(kind)
-	req.Kind = CanonicalActionKind(req.Kind)
-	fn, ok := b.Actions[kind]
-	if !ok {
-		return nil, fmt.Errorf("unknown action: %s", kind)
-	}
-	guardEnabled := b.Config == nil || b.Config.EnableActionGuards
-	checkNav := guardEnabled && shouldCheckUnexpectedNavigation(req)
-	urlReader := b.URLReader
-	if urlReader == nil {
-		urlReader = defaultActionURLReader
-		slog.Debug("URLReader is nil, using default fallback (guard checks may be no-ops without chromedp context)")
-	}
-	var beforeURL string
-	if checkNav {
-		if u, err := urlReader(ctx); err == nil {
-			beforeURL = u
-		}
-	}
-
-	res, err := fn(ctx, req)
-	if err != nil {
-		return nil, classifyActionError(err)
-	}
-
-	if checkNav && beforeURL != "" {
-		afterURL, uErr := urlReader(ctx)
-		if uErr == nil {
-			if navErr := checkUnexpectedNavigation(beforeURL, afterURL); navErr != nil {
-				return nil, navErr
-			}
-		}
-	}
-
-	return res, nil
-}
-
 // Execute delegates to TabManager.Execute for safe parallel tab execution.
 // If TabManager is not initialized, the task runs directly.
 func (b *Bridge) Execute(ctx context.Context, tabID string, task func(ctx context.Context) error) error {
@@ -721,138 +247,7 @@ func (b *Bridge) NetworkMonitor() *NetworkMonitor {
 	return b.netMonitor
 }
 
-func (b *Bridge) AvailableActions() []string {
-	keys := make([]string, 0, len(b.Actions))
-	for k := range b.Actions {
-		keys = append(keys, k)
-	}
-	return keys
-}
-
 // GetDialogManager returns the bridge's dialog manager.
 func (b *Bridge) GetDialogManager() *DialogManager {
 	return b.Dialogs
-}
-
-// ActionFunc is the type for action handlers.
-type ActionFunc func(ctx context.Context, req ActionRequest) (map[string]any, error)
-
-// ActionRequest defines the parameters for a browser action.
-//
-// Element targeting uses a unified selector string that supports multiple
-// strategies via prefix detection (see the selector package):
-//
-//	"e5"              → ref from snapshot
-//	"css:#login"      → CSS selector (explicit)
-//	"#login"          → CSS selector (auto-detected)
-//	"xpath://div"     → XPath expression
-//	"text:Submit"     → text content match
-//	"find:login btn"  → semantic / natural-language query
-//
-// For backward compatibility, the legacy Ref and Selector (CSS) fields
-// are still accepted. Call NormalizeSelector() to merge them into the
-// unified Selector field.
-type ActionRequest struct {
-	TabID    string `json:"tabId"`
-	Kind     string `json:"kind"`
-	Ref      string `json:"ref,omitempty"`
-	Selector string `json:"selector,omitempty"`
-	Text     string `json:"text"`
-	Key      string `json:"key"`
-	Value    string `json:"value"`
-	NodeID   int64  `json:"nodeId"`
-
-	// X/Y use omitempty so that re-marshaling an ActionRequest without
-	// explicit coordinates (e.g. when the tab-scoped handler forwards to
-	// the generic one) doesn't spuriously re-introduce "x":0, "y":0. The
-	// ActionRequest.UnmarshalJSON code infers HasXY from the presence of
-	// these keys, so preserving omission is what makes "use current
-	// pointer" work for mouse-down/up after a prior mouse-move. HasXY is
-	// still marshaled (with omitempty) when it was explicitly set, which
-	// preserves the explicit-click-at-(0,0) case through the round-trip.
-	X      float64 `json:"x,omitempty"`
-	Y      float64 `json:"y,omitempty"`
-	HasXY  bool    `json:"hasXY,omitempty"`
-	Button string  `json:"button,omitempty"`
-
-	ScrollX int `json:"scrollX"`
-	ScrollY int `json:"scrollY"`
-	// DeltaX/DeltaY are explicit mouse-wheel deltas for low-level
-	// mouse-wheel actions. ScrollX/ScrollY remain for backward compatibility.
-	DeltaX int `json:"deltaX,omitempty"`
-	DeltaY int `json:"deltaY,omitempty"`
-	DragX  int `json:"dragX"`
-	DragY  int `json:"dragY"`
-
-	WaitNav bool   `json:"waitNav"`
-	Fast    bool   `json:"fast"`
-	Owner   string `json:"owner"`
-
-	// DialogAction arms a one-shot dialog auto-handler before the action
-	// executes. Used when clicking a button/link that opens a JS dialog
-	// (alert/confirm/prompt). Values: "accept" or "dismiss". When set, the
-	// dialog is handled automatically without a second HTTP call.
-	DialogAction string `json:"dialogAction,omitempty"`
-	// DialogText is the optional prompt text used when DialogAction is
-	// "accept" on a prompt() dialog.
-	DialogText string `json:"dialogText,omitempty"`
-}
-
-type actionRequestAlias ActionRequest
-
-func (r *ActionRequest) UnmarshalJSON(data []byte) error {
-	var alias actionRequestAlias
-	if err := json.Unmarshal(data, &alias); err != nil {
-		return err
-	}
-	raw := map[string]json.RawMessage{}
-	if err := json.Unmarshal(data, &raw); err != nil {
-		return err
-	}
-
-	*r = ActionRequest(alias)
-	r.Kind = CanonicalActionKind(r.Kind)
-	r.HasXY = r.HasXY || hasJSONKey(raw, "x") || hasJSONKey(raw, "y")
-	if hasJSONKey(raw, "deltaX") {
-		if err := json.Unmarshal(raw["deltaX"], &r.DeltaX); err != nil {
-			return err
-		}
-	}
-	if hasJSONKey(raw, "deltaY") {
-		if err := json.Unmarshal(raw["deltaY"], &r.DeltaY); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// NormalizeSelector merges legacy Ref and Selector (CSS) fields into the
-// unified Selector field. After calling this, only Selector needs to be
-// inspected for element targeting. The method is idempotent.
-//
-// Priority: Ref > Selector (if both are set, Ref wins).
-func (r *ActionRequest) NormalizeSelector() {
-	if r.Ref != "" && r.Selector == "" {
-		// Legacy ref field → unified selector
-		r.Selector = r.Ref
-	}
-	// If Selector is already set (either from JSON or from Ref promotion),
-	// leave it as-is — Parse() will auto-detect the kind.
-}
-
-func CanonicalActionKind(kind string) string {
-	return kind
-}
-
-func hasJSONKey(raw map[string]json.RawMessage, key string) bool {
-	_, ok := raw[key]
-	return ok
-}
-
-func cryptoRandSeed() int64 {
-	n, err := rand.Int(rand.Reader, big.NewInt(1000000000))
-	if err != nil {
-		return 42
-	}
-	return n.Int64()
 }
