+++ internal/bridge/tab_manager.go
@@ -4,19 +4,15 @@ import (
 	"context"
 	"fmt"
 	"log/slog"
-	"net/url"
-	"strings"
 	"sync"
 	"time"
 
 	cdp "github.com/chromedp/cdproto/cdp"
 	"github.com/chromedp/cdproto/page"
-	"github.com/chromedp/cdproto/runtime"
 	"github.com/chromedp/cdproto/target"
 	"github.com/chromedp/chromedp"
 	"github.com/pinchtab/pinchtab/internal/config"
 	"github.com/pinchtab/pinchtab/internal/ids"
-	internalurls "github.com/pinchtab/pinchtab/internal/urls"
 )
 
 type TabSetupFunc func(ctx context.Context)
@@ -81,68 +77,9 @@ func (tm *TabManager) SetNetworkMonitor(nm *NetworkMonitor) {
 	tm.netMonitor = nm
 }
 
-func shouldBlockPopupTarget(info *target.Info) bool {
-	return info != nil && info.Type == TargetTypePage && info.OpenerID != ""
-}
-
-func (tm *TabManager) StartBrowserGuards() {
-	if tm == nil || tm.browserCtx == nil {
-		return
-	}
-
-	tm.guardOnce.Do(func() {
-		if err := chromedp.Run(tm.browserCtx, chromedp.ActionFunc(func(ctx context.Context) error {
-			c := chromedp.FromContext(ctx)
-			if c == nil || c.Browser == nil {
-				return fmt.Errorf("no browser executor")
-			}
-			return target.SetDiscoverTargets(true).Do(cdp.WithExecutor(ctx, c.Browser))
-		})); err != nil {
-			slog.Warn("browser popup guard unavailable", "err", err)
-			return
-		}
-
-		chromedp.ListenBrowser(tm.browserCtx, func(ev any) {
-			created, ok := ev.(*target.EventTargetCreated)
-			if !ok || !shouldBlockPopupTarget(created.TargetInfo) {
-				return
-			}
-
-			info := created.TargetInfo
-			go tm.closePopupTarget(info.TargetID, info.OpenerID, info.URL)
-		})
-	})
-}
-
-func (tm *TabManager) closePopupTarget(targetID, openerID target.ID, url string) {
-	closeCtx, cancel := context.WithTimeout(tm.browserCtx, 5*time.Second)
-	defer cancel()
-	logURL := internalurls.RedactForLog(url)
-
-	execCtx, err := browserExecutorContext(closeCtx)
-	if err != nil {
-		slog.Debug("popup close skipped", "targetId", targetID, "openerId", openerID, "url", logURL, "err", err)
-		return
-	}
-
-	if err := target.CloseTarget(targetID).Do(execCtx); err != nil {
-		slog.Debug("popup close failed", "targetId", targetID, "openerId", openerID, "url", logURL, "err", err)
-		return
-	}
-
-	slog.Info("blocked popup target", "targetId", targetID, "openerId", openerID, "url", logURL)
-}
-
-func (tm *TabManager) markAccessed(tabID string) {
-	tm.mu.Lock()
-	tm.accessed[tabID] = true
-	if entry, ok := tm.tabs[tabID]; ok {
-		entry.LastUsed = time.Now()
-	}
-	tm.currentTab = tabID
-	tm.mu.Unlock()
-}
-
+// browserExecutorContext returns a context bound to the top-level browser
+// executor, suitable for issuing browser-scoped CDP calls (e.g. target.*).
+// Shared helper used by tab lifecycle, lookup, popup-guard, and cleanup paths.
 func browserExecutorContext(ctx context.Context) (context.Context, error) {
 	if ctx == nil {
 		return nil, fmt.Errorf("no browser context available")
@@ -152,169 +89,6 @@ func browserExecutorContext(ctx context.Context) (context.Context, error) {
 		return nil, fmt.Errorf("no browser executor available")
 	}
 	return cdp.WithExecutor(ctx, c.Browser), nil
-}
-
-// selectCurrentTrackedTab returns the current tab ID, falling back to the most
-// recently used tab if the explicit pointer is stale or unset.
-func (tm *TabManager) selectCurrentTrackedTab() string {
-	// Prefer explicit current tab if still tracked
-	if tm.currentTab != "" {
-		if _, ok := tm.tabs[tm.currentTab]; ok {
-			return tm.currentTab
-		}
-	}
-
-	// Fallback: pick the tab with the most recent LastUsed
-	var best string
-	var bestTime time.Time
-	for id, entry := range tm.tabs {
-		if entry.LastUsed.After(bestTime) {
-			best = id
-			bestTime = entry.LastUsed
-		}
-	}
-	// If no LastUsed set, fall back to most recent CreatedAt
-	if best == "" {
-		for id, entry := range tm.tabs {
-			if entry.CreatedAt.After(bestTime) {
-				best = id
-				bestTime = entry.CreatedAt
-			}
-		}
-	}
-	return best
-}
-
-// AccessedTabIDs returns the set of tab IDs that were accessed this session.
-func (tm *TabManager) AccessedTabIDs() map[string]bool {
-	tm.mu.RLock()
-	defer tm.mu.RUnlock()
-	out := make(map[string]bool, len(tm.accessed))
-	for k := range tm.accessed {
-		out[k] = true
-	}
-	return out
-}
-
-func (tm *TabManager) TabContext(tabID string) (context.Context, string, error) {
-	if tm == nil {
-		return nil, "", fmt.Errorf("tab manager not initialized")
-	}
-	if tabID == "" {
-		// Resolve to current tracked tab
-		tm.mu.RLock()
-		tabID = tm.selectCurrentTrackedTab()
-		tm.mu.RUnlock()
-
-		if tabID == "" {
-			// No tracked tabs — try to find one from CDP targets
-			targets, err := tm.ListTargets()
-			if err != nil {
-				return nil, "", fmt.Errorf("list targets: %w", err)
-			}
-			if len(targets) == 0 {
-				return nil, "", fmt.Errorf("no tabs open")
-			}
-			rawID := string(targets[0].TargetID)
-			tabID = tm.idMgr.TabIDFromCDPTarget(rawID)
-		}
-	}
-
-	tm.mu.RLock()
-	entry, ok := tm.tabs[tabID]
-	tm.mu.RUnlock()
-
-	if !ok {
-		// Attempt to auto-track the tab if it's open but untracked
-		targets, err := tm.ListTargets()
-		if err == nil {
-			for _, t := range targets {
-				raw := string(t.TargetID)
-				if tm.idMgr.TabIDFromCDPTarget(raw) == tabID {
-					// Initialize context and register it
-					ctx, cancel := chromedp.NewContext(tm.browserCtx, chromedp.WithTargetID(target.ID(raw)))
-					if tm.onTabSetup != nil {
-						tm.onTabSetup(ctx)
-					}
-					if tm.netMonitor != nil {
-						if err := tm.netMonitor.StartCapture(ctx, tabID); err != nil {
-							slog.Warn("eager network capture failed", "tab", tabID, "err", err)
-						}
-					}
-					if tm.dialogMgr != nil {
-						autoAccept := tm.config != nil && tm.config.DialogAutoAccept
-						ListenDialogEvents(ctx, tabID, tm.dialogMgr, autoAccept)
-						if err := EnableDialogEvents(ctx); err != nil {
-							slog.Warn("enable dialog events failed", "tabId", tabID, "err", err)
-						}
-					}
-					tm.RegisterTabWithCancel(tabID, raw, ctx, cancel)
-
-					tm.mu.RLock()
-					entry = tm.tabs[tabID]
-					tm.mu.RUnlock()
-					ok = true
-					break
-				}
-			}
-		}
-	}
-
-	if !ok {
-		return nil, "", fmt.Errorf("tab %s not found", tabID)
-	}
-
-	if entry.Ctx == nil {
-		return nil, "", fmt.Errorf("tab %s has no active context", tabID)
-	}
-
-	tm.markAccessed(tabID)
-
-	return entry.Ctx, tabID, nil
-}
-
-// closeOldestTab evicts the tab with the earliest CreatedAt timestamp.
-func (tm *TabManager) closeOldestTab() error {
-	tm.mu.RLock()
-	var oldestID string
-	var oldestTime time.Time
-	for id, entry := range tm.tabs {
-		if oldestID == "" || entry.CreatedAt.Before(oldestTime) {
-			oldestID = id
-			oldestTime = entry.CreatedAt
-		}
-	}
-	tm.mu.RUnlock()
-
-	if oldestID == "" {
-		return fmt.Errorf("no tabs to evict")
-	}
-	slog.Info("evicting oldest tab", "id", oldestID, "createdAt", oldestTime)
-	return tm.CloseTab(oldestID)
-}
-
-// closeLRUTab evicts the tab with the earliest LastUsed timestamp.
-func (tm *TabManager) closeLRUTab() error {
-	tm.mu.RLock()
-	var lruID string
-	var lruTime time.Time
-	for id, entry := range tm.tabs {
-		t := entry.LastUsed
-		if t.IsZero() {
-			t = entry.CreatedAt
-		}
-		if lruID == "" || t.Before(lruTime) {
-			lruID = id
-			lruTime = t
-		}
-	}
-	tm.mu.RUnlock()
-
-	if lruID == "" {
-		return fmt.Errorf("no tabs to evict")
-	}
-	slog.Info("evicting LRU tab", "id", lruID, "lastUsed", lruTime)
-	return tm.CloseTab(lruID)
 }
 
 func (tm *TabManager) CreateTab(url string) (string, context.Context, context.CancelFunc, error) {
@@ -365,7 +139,6 @@ func (tm *TabManager) CreateTab(url string) (string, context.Context, context.Ca
 	}
 	createCancel()
 
-	// Create a context for the new tab
 	ctx, cancel := chromedp.NewContext(tm.browserCtx,
 		chromedp.WithTargetID(targetID),
 	)
@@ -415,7 +188,6 @@ func (tm *TabManager) CreateTab(url string) (string, context.Context, context.Ca
 
 	now := time.Now()
 
-	// Set up dialog event listening for this tab
 	if tm.dialogMgr != nil {
 		autoAccept := tm.config != nil && tm.config.DialogAutoAccept
 		ListenDialogEvents(ctx, tabID, tm.dialogMgr, autoAccept)
@@ -469,7 +241,6 @@ func (tm *TabManager) CloseTab(tabID string) error {
 		entry.Cancel()
 	}
 
-	// Resolve to raw CDP target ID for the actual CDP close call
 	cdpTargetID := tabID
 	if tracked && entry.CDPID != "" {
 		cdpTargetID = entry.CDPID
@@ -509,7 +280,6 @@ func (tm *TabManager) FocusTab(tabID string) error {
 		return err
 	}
 
-	// Bring the tab to front via CDP
 	if err := chromedp.Run(ctx, chromedp.ActionFunc(func(ctx context.Context) error {
 		return page.BringToFront().Do(ctx)
 	})); err != nil {
@@ -526,152 +296,6 @@ func (tm *TabManager) FocusTab(tabID string) error {
 	return nil
 }
 
-// ResolveTabByIndex resolves a 1-based tab index to a tab ID.
-// Returns the tab ID and its URL/title for display.
-func (tm *TabManager) ResolveTabByIndex(index int) (string, string, string, error) {
-	targets, err := tm.ListTargets()
-	if err != nil {
-		return "", "", "", err
-	}
-	if index < 1 || index > len(targets) {
-		return "", "", "", fmt.Errorf("tab index %d out of range (1-%d)", index, len(targets))
-	}
-	t := targets[index-1]
-	tabID := tm.idMgr.TabIDFromCDPTarget(string(t.TargetID))
-	return tabID, t.URL, t.Title, nil
-}
-
-func (tm *TabManager) ListTargets() ([]*target.Info, error) {
-	if tm == nil {
-		return nil, fmt.Errorf("tab manager not initialized")
-	}
-	if tm.browserCtx == nil {
-		return nil, fmt.Errorf("no browser connection")
-	}
-
-	execCtx, err := browserExecutorContext(tm.browserCtx)
-	if err != nil {
-		return nil, fmt.Errorf("browser executor: %w", err)
-	}
-
-	ctx, cancel := context.WithTimeout(execCtx, 5*time.Second)
-	defer cancel()
-
-	targets, err := target.GetTargets().Do(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("get targets: %w", err)
-	}
-
-	pages := make([]*target.Info, 0, len(targets))
-	for _, t := range targets {
-		if t.Type == TargetTypePage {
-			pages = append(pages, t)
-		}
-	}
-	return pages, nil
-}
-
-// ListTargetsWithContext is like ListTargets but uses a custom context for timeout control.
-func (tm *TabManager) ListTargetsWithContext(ctx context.Context) ([]*target.Info, error) {
-	if tm == nil {
-		return nil, fmt.Errorf("tab manager not initialized")
-	}
-	if tm.browserCtx == nil {
-		return nil, fmt.Errorf("no browser connection")
-	}
-
-	execCtx, err := browserExecutorContext(tm.browserCtx)
-	if err != nil {
-		return nil, fmt.Errorf("browser executor: %w", err)
-	}
-
-	mergedCtx, cancel := context.WithCancel(execCtx)
-	defer cancel()
-
-	go func() {
-		select {
-		case <-ctx.Done():
-			cancel()
-		case <-mergedCtx.Done():
-		}
-	}()
-
-	targets, err := target.GetTargets().Do(mergedCtx)
-	if err != nil {
-		return nil, fmt.Errorf("get targets: %w", err)
-	}
-
-	pages := make([]*target.Info, 0, len(targets))
-	for _, t := range targets {
-		if t.Type == TargetTypePage {
-			pages = append(pages, t)
-		}
-	}
-	return pages, nil
-}
-
-func (tm *TabManager) GetRefCache(tabID string) *RefCache {
-	tm.mu.RLock()
-	defer tm.mu.RUnlock()
-	return tm.snapshots[tabID]
-}
-
-func (tm *TabManager) SetRefCache(tabID string, cache *RefCache) {
-	tm.mu.Lock()
-	defer tm.mu.Unlock()
-	tm.snapshots[tabID] = cache
-}
-
-func (tm *TabManager) DeleteRefCache(tabID string) {
-	tm.mu.Lock()
-	defer tm.mu.Unlock()
-	delete(tm.snapshots, tabID)
-}
-
-func (tm *TabManager) GetFrameScope(tabID string) (FrameScope, bool) {
-	tm.mu.RLock()
-	defer tm.mu.RUnlock()
-	scope, ok := tm.frameScope[tabID]
-	return scope, ok && scope.Active()
-}
-
-func (tm *TabManager) SetFrameScope(tabID string, scope FrameScope) {
-	tm.mu.Lock()
-	defer tm.mu.Unlock()
-	if !scope.Active() {
-		delete(tm.frameScope, tabID)
-		return
-	}
-	tm.frameScope[tabID] = scope
-}
-
-func (tm *TabManager) ClearFrameScope(tabID string) {
-	tm.mu.Lock()
-	defer tm.mu.Unlock()
-	delete(tm.frameScope, tabID)
-}
-
-func (tm *TabManager) RegisterTab(tabID string, ctx context.Context) {
-	now := time.Now()
-	tm.mu.Lock()
-	tm.tabs[tabID] = &TabEntry{Ctx: ctx, CreatedAt: now, LastUsed: now}
-	tm.currentTab = tabID
-	tm.mu.Unlock()
-
-	tm.startTabPolicyWatcher(tabID, ctx)
-}
-
-// RegisterTabWithCancel registers a tab ID with its context and cancel function.
-func (tm *TabManager) RegisterTabWithCancel(tabID, rawCDPID string, ctx context.Context, cancel context.CancelFunc) {
-	now := time.Now()
-	tm.mu.Lock()
-	tm.tabs[tabID] = &TabEntry{Ctx: ctx, Cancel: cancel, CDPID: rawCDPID, CreatedAt: now, LastUsed: now}
-	tm.currentTab = tabID
-	tm.mu.Unlock()
-
-	tm.startTabPolicyWatcher(tabID, ctx)
-}
-
 // Execute runs a task for a tab through the TabExecutor, ensuring per-tab
 // sequential execution with cross-tab parallelism bounded by the semaphore.
 // If the TabExecutor has not been initialized, the task runs directly.
@@ -685,344 +309,4 @@ func (tm *TabManager) Execute(ctx context.Context, tabID string, task func(ctx c
 // Executor returns the underlying TabExecutor (may be nil).
 func (tm *TabManager) Executor() *TabExecutor {
 	return tm.executor
-}
-
-func (tm *TabManager) CleanStaleTabs(ctx context.Context, interval time.Duration) {
-	ticker := time.NewTicker(interval)
-	defer ticker.Stop()
-
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-ticker.C:
-		}
-
-		targets, err := tm.ListTargets()
-		if err != nil {
-			continue
-		}
-
-		alive := make(map[string]bool, len(targets))
-		for _, t := range targets {
-			alive[string(t.TargetID)] = true
-		}
-
-		type staleTab struct {
-			tabID string
-			cdpID string
-		}
-		var staleTabs []staleTab
-		tm.mu.RLock()
-		for id, entry := range tm.tabs {
-			if !alive[id] {
-				cdpID := entry.CDPID
-				if cdpID == "" {
-					cdpID = id
-				}
-				staleTabs = append(staleTabs, staleTab{tabID: id, cdpID: cdpID})
-			}
-		}
-		tm.mu.RUnlock()
-
-		for _, stale := range staleTabs {
-			tm.purgeTrackedTabState(stale.tabID, stale.cdpID)
-			slog.Info("cleaned stale tab", "id", stale.tabID)
-		}
-	}
-}
-
-func (tm *TabManager) purgeTrackedTabState(tabID, cdpTargetID string) bool {
-	resolvedTabID, resolvedCDPID, cancel, ok := tm.lookupTrackedTabForCleanup(tabID, cdpTargetID)
-	if !ok {
-		return false
-	}
-	if cancel != nil {
-		cancel()
-	}
-
-	tm.mu.Lock()
-	if entry, ok := tm.tabs[resolvedTabID]; ok && entry.autoCloseTimer != nil {
-		entry.autoCloseTimer.Stop()
-		entry.autoCloseTimer = nil
-		entry.autoCloseGen++
-	}
-	delete(tm.tabs, resolvedTabID)
-	delete(tm.snapshots, resolvedTabID)
-	delete(tm.frameScope, resolvedTabID)
-	delete(tm.accessed, resolvedTabID)
-	if tm.currentTab == resolvedTabID {
-		tm.currentTab = ""
-	}
-	tm.mu.Unlock()
-
-	if tm.dialogMgr != nil {
-		tm.dialogMgr.ClearPending(resolvedTabID)
-	}
-	if tm.executor != nil {
-		tm.executor.RemoveTab(resolvedTabID)
-	}
-	if tm.logStore != nil {
-		tm.logStore.RemoveTab(resolvedCDPID)
-	}
-	// Notify listeners (e.g. session persistence) that a tab disappeared,
-	// regardless of whether the trigger was a deliberate CloseTab, an eviction,
-	// the auto-close lifecycle timer, or Chrome reporting the target gone
-	// (CleanStaleTabs / user closing the tab in headed Chrome directly).
-	if tm.onAfterClose != nil {
-		tm.onAfterClose()
-	}
-	return true
-}
-
-func (tm *TabManager) lookupTrackedTabForCleanup(tabID, cdpTargetID string) (string, string, context.CancelFunc, bool) {
-	tm.mu.RLock()
-	defer tm.mu.RUnlock()
-
-	if tabID != "" {
-		if entry, ok := tm.tabs[tabID]; ok {
-			resolvedCDPID := cdpTargetID
-			if resolvedCDPID == "" {
-				resolvedCDPID = entry.CDPID
-			}
-			if resolvedCDPID == "" {
-				resolvedCDPID = tabID
-			}
-			return tabID, resolvedCDPID, entry.Cancel, true
-		}
-	}
-
-	if cdpTargetID == "" {
-		return "", "", nil, false
-	}
-
-	for id, entry := range tm.tabs {
-		resolvedCDPID := entry.CDPID
-		if resolvedCDPID == "" {
-			resolvedCDPID = id
-		}
-		if id == cdpTargetID || resolvedCDPID == cdpTargetID {
-			return id, resolvedCDPID, entry.Cancel, true
-		}
-	}
-	return "", "", nil, false
-}
-
-func (tm *TabManager) purgeTrackedTabStateByTargetID(cdpTargetID string) bool {
-	return tm.purgeTrackedTabState("", cdpTargetID)
-}
-
-// setupConsoleCapture enables runtime domain and listens for console/exception events.
-func (tm *TabManager) setupConsoleCapture(ctx context.Context, rawCDPID string) {
-	if tm.logStore == nil {
-		return
-	}
-
-	execContextSources := make(map[runtime.ExecutionContextID]string)
-	var execContextMu sync.RWMutex
-
-	// Listen for console API calls and exceptions
-	chromedp.ListenTarget(ctx, func(ev any) {
-		switch ev := ev.(type) {
-		case *runtime.EventExecutionContextCreated:
-			if ev.Context == nil {
-				return
-			}
-			execContextMu.Lock()
-			execContextSources[ev.Context.ID] = executionContextSource(ev.Context)
-			execContextMu.Unlock()
-
-		case *runtime.EventExecutionContextDestroyed:
-			execContextMu.Lock()
-			delete(execContextSources, ev.ExecutionContextID)
-			execContextMu.Unlock()
-
-		case *runtime.EventExecutionContextsCleared:
-			execContextMu.Lock()
-			clear(execContextSources)
-			execContextMu.Unlock()
-
-		case *runtime.EventConsoleAPICalled:
-			var msg string
-			for _, arg := range ev.Args {
-				if len(arg.Value) > 0 {
-					// arg.Value is jsontext.Value ([]byte), use as string directly
-					// Strip quotes if it's a JSON string
-					val := string(arg.Value)
-					if len(val) >= 2 && val[0] == '"' && val[len(val)-1] == '"' {
-						val = val[1 : len(val)-1]
-					}
-					msg += val
-				} else if arg.Description != "" {
-					msg += arg.Description
-				}
-				msg += " "
-			}
-
-			var ts time.Time
-			if ev.Timestamp != nil {
-				ts = time.Time(*ev.Timestamp)
-			} else {
-				ts = time.Now()
-			}
-
-			source := stackTraceSource(ev.StackTrace)
-			if source == "" {
-				execContextMu.RLock()
-				source = execContextSources[ev.ExecutionContextID]
-				execContextMu.RUnlock()
-			}
-			if source == "" {
-				source = strings.TrimSpace(ev.Context)
-			}
-			if isInternalConsoleSource(source) {
-				return
-			}
-
-			tm.logStore.AddConsoleLog(rawCDPID, LogEntry{
-				Timestamp: ts,
-				Level:     string(ev.Type),
-				Message:   msg,
-				Source:    source,
-			})
-
-		case *runtime.EventExceptionThrown:
-			msg := ev.ExceptionDetails.Text
-			if ev.ExceptionDetails.Exception != nil && ev.ExceptionDetails.Exception.Description != "" {
-				msg += ": " + ev.ExceptionDetails.Exception.Description
-			}
-
-			var ts time.Time
-			if ev.Timestamp != nil {
-				ts = time.Time(*ev.Timestamp)
-			} else {
-				ts = time.Now()
-			}
-
-			source := exceptionSource(ev.ExceptionDetails)
-			if source == "" {
-				execContextMu.RLock()
-				source = execContextSources[ev.ExceptionDetails.ExecutionContextID]
-				execContextMu.RUnlock()
-			}
-			if isInternalConsoleSource(source) {
-				return
-			}
-
-			stack := ""
-			if ev.ExceptionDetails.Exception != nil {
-				stack = ev.ExceptionDetails.Exception.Description
-			}
-
-			tm.logStore.AddErrorLog(rawCDPID, ErrorEntry{
-				Timestamp: ts,
-				Message:   msg,
-				URL:       ev.ExceptionDetails.URL,
-				Line:      ev.ExceptionDetails.LineNumber,
-				Column:    ev.ExceptionDetails.ColumnNumber,
-				Stack:     stack,
-			})
-		}
-	})
-
-	// Enable runtime domain to receive console/exception events
-	go func() {
-		_ = chromedp.Run(ctx, chromedp.ActionFunc(func(c context.Context) error {
-			return runtime.Enable().Do(c)
-		}))
-	}()
-}
-
-func (tm *TabManager) shouldEagerlyCaptureConsole() bool {
-	if tm == nil || tm.config == nil {
-		return true
-	}
-	return !strings.EqualFold(strings.TrimSpace(tm.config.StealthLevel), "full")
-}
-
-func (tm *TabManager) EnsureConsoleCapture(tabID string) {
-	if tm == nil || tm.logStore == nil {
-		return
-	}
-
-	tm.mu.Lock()
-	entry := tm.tabs[tabID]
-	if entry == nil && tabID == "" {
-		entry = tm.tabs[tm.currentTab]
-	}
-	if entry == nil || entry.Ctx == nil || entry.ConsoleCaptureEnabled {
-		tm.mu.Unlock()
-		return
-	}
-	entry.ConsoleCaptureEnabled = true
-	ctx := entry.Ctx
-	rawCDPID := entry.CDPID
-	tm.mu.Unlock()
-
-	tm.setupConsoleCapture(ctx, rawCDPID)
-}
-
-func executionContextSource(ctx *runtime.ExecutionContextDescription) string {
-	if ctx == nil {
-		return ""
-	}
-	if source := strings.TrimSpace(ctx.Origin); source != "" {
-		return source
-	}
-	return strings.TrimSpace(ctx.Name)
-}
-
-func exceptionSource(details *runtime.ExceptionDetails) string {
-	if details == nil {
-		return ""
-	}
-	if source := strings.TrimSpace(details.URL); source != "" {
-		return source
-	}
-	return stackTraceSource(details.StackTrace)
-}
-
-func stackTraceSource(trace *runtime.StackTrace) string {
-	for trace != nil {
-		for _, frame := range trace.CallFrames {
-			if frame == nil {
-				continue
-			}
-			if source := strings.TrimSpace(frame.URL); source != "" {
-				return source
-			}
-		}
-		trace = trace.Parent
-	}
-	return ""
-}
-
-func isInternalConsoleSource(source string) bool {
-	source = strings.TrimSpace(source)
-	if source == "" {
-		return false
-	}
-
-	lower := strings.ToLower(source)
-	switch {
-	case strings.HasPrefix(lower, "chrome-extension://"),
-		strings.HasPrefix(lower, "edge-extension://"),
-		strings.HasPrefix(lower, "moz-extension://"),
-		strings.HasPrefix(lower, "safari-extension://"),
-		strings.HasPrefix(lower, "devtools://"),
-		strings.HasPrefix(lower, "chrome://"),
-		strings.HasPrefix(lower, "about:"):
-		return true
-	}
-
-	parsed, err := url.Parse(source)
-	if err != nil {
-		return false
-	}
-	switch strings.ToLower(parsed.Scheme) {
-	case "chrome-extension", "edge-extension", "moz-extension", "safari-extension", "devtools", "chrome", "about":
-		return true
-	default:
-		return false
-	}
 }
