+++ internal/handlers/download.go
@@ -1,24 +1,17 @@
 package handlers
 
 import (
-	"compress/gzip"
 	"context"
 	"encoding/base64"
-	"errors"
 	"fmt"
-	"io"
 	"log/slog"
 	"net/http"
 	"net/url"
 	"os"
-	"path"
 	"path/filepath"
-	"strconv"
 	"strings"
-	"sync"
-	"time"
-
 	"sync/atomic"
+	"time"
 
 	"github.com/chromedp/cdproto/cdp"
 	"github.com/chromedp/cdproto/fetch"
@@ -26,225 +19,8 @@ import (
 	"github.com/chromedp/chromedp"
 	"github.com/pinchtab/pinchtab/internal/activity"
 	"github.com/pinchtab/pinchtab/internal/authn"
-	"github.com/pinchtab/pinchtab/internal/bridge"
-	"github.com/pinchtab/pinchtab/internal/config"
 	"github.com/pinchtab/pinchtab/internal/httpx"
-	"github.com/pinchtab/pinchtab/internal/idpi"
-	"github.com/pinchtab/pinchtab/internal/netguard"
 )
-
-var errDownloadTooLarge = errors.New("download response too large")
-
-type downloadURLGuard struct {
-	allowedDomains []string
-}
-
-func newDownloadURLGuard(allowedDomains []string) *downloadURLGuard {
-	return &downloadURLGuard{allowedDomains: append([]string(nil), allowedDomains...)}
-}
-
-func (g *downloadURLGuard) isHostAllowed(host string) bool {
-	if len(g.allowedDomains) == 0 {
-		return false
-	}
-	host = netguard.NormalizeHost(host)
-	if host == "" {
-		return false
-	}
-	return g.isDomainAllowed("https://" + host)
-}
-
-// isDomainAllowed reports whether rawURL's domain is on the configured
-// allowlist. Allowlisted domains bypass private-IP checks because they
-// are explicitly trusted by the operator (e.g. internal docker hosts).
-func (g *downloadURLGuard) isDomainAllowed(rawURL string) bool {
-	if len(g.allowedDomains) == 0 {
-		return false
-	}
-	result := idpi.CheckDomain(rawURL, config.IDPIConfig{
-		Enabled:    true,
-		StrictMode: true,
-	}, g.allowedDomains)
-	return !result.Blocked
-}
-
-func (g *downloadURLGuard) Validate(rawURL string) error {
-	parsed, err := url.Parse(rawURL)
-	if err != nil {
-		return fmt.Errorf("invalid URL")
-	}
-
-	if parsed.Scheme != "http" && parsed.Scheme != "https" {
-		return fmt.Errorf("only http/https schemes are allowed")
-	}
-
-	host := netguard.NormalizeHost(parsed.Hostname())
-	if host == "" || netguard.IsLocalHost(host) {
-		return fmt.Errorf("internal or blocked host")
-	}
-
-	// Allowlisted domains bypass IP validation (e.g. internal docker hosts).
-	if len(g.allowedDomains) > 0 {
-		result := idpi.CheckDomain(rawURL, config.IDPIConfig{
-			Enabled:    true,
-			StrictMode: true,
-		}, g.allowedDomains)
-		if result.Blocked {
-			return fmt.Errorf("domain not allowed by security.downloadAllowedDomains")
-		}
-		return nil
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 8*time.Second)
-	defer cancel()
-
-	if _, err := netguard.ResolveAndValidatePublicIPs(ctx, host); err != nil {
-		if errors.Is(err, netguard.ErrResolveHost) {
-			return fmt.Errorf("could not resolve host")
-		}
-		if errors.Is(err, netguard.ErrPrivateInternalIP) {
-			return fmt.Errorf("private/internal IP blocked")
-		}
-		return fmt.Errorf("could not resolve host")
-	}
-	return nil
-}
-
-func validateDownloadRemoteIPAddress(raw string) error {
-	raw = strings.TrimSpace(raw)
-	if raw == "" {
-		// Best-effort mitigation: some responses may not expose a remote IP
-		// (for example, cached responses). Skip the post-connect check then.
-		return nil
-	}
-
-	normalized := netguard.NormalizeRemoteIP(raw)
-	if err := netguard.ValidateRemoteIPAddress(raw); err != nil {
-		if errors.Is(err, netguard.ErrUnparseableRemoteIP) {
-			return fmt.Errorf("download connected to an unparseable remote IP %q", normalized)
-		}
-		if errors.Is(err, netguard.ErrPrivateInternalIP) {
-			return fmt.Errorf("download connected to blocked remote IP %s", normalized)
-		}
-		return fmt.Errorf("download connected to an unparseable remote IP %q", raw)
-	}
-	return nil
-}
-
-// validateDownloadURL blocks file://, internal hosts, private IPs, and cloud metadata.
-// Only public http/https URLs are allowed.
-func validateDownloadURL(rawURL string) error {
-	return newDownloadURLGuard(nil).Validate(rawURL)
-}
-
-func validateTabScopedDownloadURL(currentURL, requestedURL string) error {
-	currentParsed, err := url.Parse(strings.TrimSpace(currentURL))
-	if err != nil || currentParsed.Scheme == "" || currentParsed.Host == "" {
-		return fmt.Errorf("tab-scoped downloads require an active http(s) page")
-	}
-	if currentParsed.Scheme != "http" && currentParsed.Scheme != "https" {
-		return fmt.Errorf("tab-scoped downloads require an active http(s) page")
-	}
-
-	requestedParsed, err := url.Parse(strings.TrimSpace(requestedURL))
-	if err != nil || requestedParsed.Scheme == "" || requestedParsed.Host == "" {
-		return fmt.Errorf("invalid download URL")
-	}
-
-	if strings.EqualFold(currentParsed.Scheme, requestedParsed.Scheme) &&
-		strings.EqualFold(currentParsed.Host, requestedParsed.Host) {
-		return nil
-	}
-
-	return fmt.Errorf("tab-scoped downloads are limited to the current page origin")
-}
-
-type downloadRequestGuard struct {
-	validator    *downloadURLGuard
-	maxRedirects int
-	redirects    atomic.Int32
-
-	mu         sync.Mutex
-	blockedErr error
-}
-
-func newDownloadRequestGuard(validator *downloadURLGuard, maxRedirects int) *downloadRequestGuard {
-	return &downloadRequestGuard{
-		validator:    validator,
-		maxRedirects: maxRedirects,
-	}
-}
-
-func (g *downloadRequestGuard) Validate(rawURL string, redirected bool) error {
-	// Skip validation for Chrome internal URLs (about:blank, chrome-error://, etc.)
-	// that fire during tab creation before the actual navigation begins.
-	if rawURL == "about:blank" || strings.HasPrefix(rawURL, "chrome") {
-		return nil
-	}
-	if err := g.validator.Validate(rawURL); err != nil {
-		return fmt.Errorf("unsafe browser request: %w", err)
-	}
-	if redirected && g.maxRedirects >= 0 {
-		count := int(g.redirects.Add(1))
-		if count > g.maxRedirects {
-			return fmt.Errorf("%w: got %d, max %d", bridge.ErrTooManyRedirects, count, g.maxRedirects)
-		}
-	}
-	return nil
-}
-
-func (g *downloadRequestGuard) NoteBlocked(err error) {
-	g.mu.Lock()
-	if g.blockedErr == nil {
-		g.blockedErr = err
-	}
-	g.mu.Unlock()
-}
-
-func (g *downloadRequestGuard) BlockedError() error {
-	g.mu.Lock()
-	defer g.mu.Unlock()
-	return g.blockedErr
-}
-
-func downloadTooLargeError(size int64, maxBytes int) error {
-	return fmt.Errorf("%w: received %d bytes, max %d", errDownloadTooLarge, size, maxBytes)
-}
-
-func parseContentLengthHeader(headers network.Headers) (int64, bool) {
-	for key, raw := range headers {
-		if !strings.EqualFold(strings.TrimSpace(key), "Content-Length") {
-			continue
-		}
-		value := strings.TrimSpace(fmt.Sprint(raw))
-		if value == "" {
-			return 0, false
-		}
-		size, err := strconv.ParseInt(value, 10, 64)
-		if err != nil || size < 0 {
-			return 0, false
-		}
-		return size, true
-	}
-	return 0, false
-}
-
-func writeDownloadGuardError(w http.ResponseWriter, err error, maxBytes int) bool {
-	if err == nil {
-		return false
-	}
-	switch {
-	case errors.Is(err, bridge.ErrTooManyRedirects):
-		httpx.Error(w, 422, fmt.Errorf("download: %w", err))
-	case errors.Is(err, errDownloadTooLarge):
-		httpx.ErrorCode(w, http.StatusRequestEntityTooLarge, "download_too_large", err.Error(), false, map[string]any{
-			"maxBytes": maxBytes,
-		})
-	default:
-		httpx.Error(w, 400, err)
-	}
-	return true
-}
 
 // HandleDownload fetches a URL using the browser's session (cookies, stealth)
 // and returns the content. This preserves authentication and fingerprint.
@@ -332,7 +108,6 @@ func (h *Handlers) HandleDownload(w http.ResponseWriter, r *http.Request) {
 	defer tCancel()
 	go httpx.CancelOnClientDone(r.Context(), tCancel)
 
-	// Enable network tracking to capture response metadata.
 	var requestID network.RequestID
 	var responseMIME string
 	var responseStatus int
@@ -464,7 +239,6 @@ func (h *Handlers) HandleDownload(w http.ResponseWriter, r *http.Request) {
 			slog.Info("download: Chrome navigation aborted, falling back to direct fetch", "url", dlURL)
 			body, mime, status, fetchErr := h.fetchDirectWithCookies(tCtx, browserCtx, dlURL, validator, maxDownloadBytes)
 			if fetchErr != nil {
-				// Return 400 for redirect-blocked errors (SSRF protection)
 				errMsg := fetchErr.Error()
 				if strings.Contains(errMsg, "blocked") || strings.Contains(errMsg, "private") {
 					httpx.Error(w, 400, fmt.Errorf("unsafe browser request: %w", fetchErr))
@@ -486,7 +260,6 @@ func (h *Handlers) HandleDownload(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	// Wait for response.
 	select {
 	case <-done:
 	case <-tCtx.Done():
@@ -510,7 +283,6 @@ func (h *Handlers) HandleDownload(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	// Get response body via CDP.
 	var body []byte
 	if err := chromedp.Run(tCtx,
 		chromedp.ActionFunc(func(ctx context.Context) error {
@@ -590,113 +362,6 @@ func (h *Handlers) writeDownloadResponse(w http.ResponseWriter, body []byte, mim
 		"size":        len(body),
 		"url":         dlURL,
 	})
-}
-
-// fetchDirectWithCookies performs a Go HTTP fetch with browser cookies.
-// Fallback for when Chrome navigation aborts (e.g. .gz files).
-func (h *Handlers) fetchDirectWithCookies(ctx context.Context, browserCtx context.Context, dlURL string, validator *downloadURLGuard, maxBytes int) (body []byte, contentType string, statusCode int, err error) {
-	var browserCookies []*network.Cookie
-	if fetchErr := chromedp.Run(browserCtx,
-		chromedp.ActionFunc(func(ctx context.Context) error {
-			var err error
-			browserCookies, err = network.GetCookies().WithURLs([]string{dlURL}).Do(ctx)
-			return err
-		}),
-	); fetchErr != nil {
-		slog.Debug("download fallback: failed to get browser cookies", "err", fetchErr)
-	}
-
-	req, err := http.NewRequestWithContext(ctx, "GET", dlURL, nil)
-	if err != nil {
-		return nil, "", 0, err
-	}
-	if h.Config.UserAgent != "" {
-		req.Header.Set("User-Agent", h.Config.UserAgent)
-	}
-	req.Header.Set("Accept", "*/*")
-
-	for _, c := range browserCookies {
-		req.AddCookie(&http.Cookie{Name: c.Name, Value: c.Value})
-	}
-
-	client := newGuardedDownloadClient(validator, h.Config.MaxRedirects, 30)
-	resp, err := client.Do(req)
-	if err != nil {
-		return nil, "", 0, err
-	}
-	defer func() { _ = resp.Body.Close() }()
-
-	if cl := resp.Header.Get("Content-Length"); cl != "" {
-		if contentLength, parseErr := strconv.ParseInt(cl, 10, 64); parseErr == nil {
-			if contentLength > int64(maxBytes) {
-				return nil, "", 0, errDownloadTooLarge
-			}
-		}
-	}
-
-	reader := io.Reader(resp.Body)
-	isGzip := isGzipContent(resp.Header.Get("Content-Type"), dlURL) &&
-		!strings.EqualFold(resp.Header.Get("Content-Encoding"), "gzip")
-
-	if isGzip {
-		gz, gzErr := gzip.NewReader(resp.Body)
-		if gzErr != nil {
-			return nil, "", 0, fmt.Errorf("gzip decompress: %w", gzErr)
-		}
-		defer func() { _ = gz.Close() }()
-		reader = gz
-	}
-
-	// LimitReader on decompressed stream protects against gzip bombs.
-	data, err := io.ReadAll(io.LimitReader(reader, int64(maxBytes)+1))
-	if err != nil {
-		return nil, "", 0, err
-	}
-	if len(data) > maxBytes {
-		return nil, "", 0, errDownloadTooLarge
-	}
-
-	ct := resp.Header.Get("Content-Type")
-	if ct == "" {
-		ct = "application/octet-stream"
-	}
-	// If we decompressed gzip, report the inner content type
-	if isGzip {
-		ct = inferDecompressedContentType(dlURL, ct)
-	}
-
-	return data, ct, resp.StatusCode, nil
-}
-
-func isGzipContent(contentType, rawURL string) bool {
-	if strings.Contains(strings.ToLower(contentType), "gzip") {
-		return true
-	}
-	// Use path.Ext for precise extension matching (.gz, not .pgz or .ngz)
-	parsed, err := url.Parse(rawURL)
-	if err != nil {
-		return false
-	}
-	return strings.EqualFold(path.Ext(parsed.Path), ".gz")
-}
-
-func inferDecompressedContentType(rawURL, fallback string) string {
-	lower := strings.ToLower(rawURL)
-	// Strip .gz to infer inner type
-	if strings.HasSuffix(lower, ".xml.gz") {
-		return "application/xml"
-	}
-	if strings.HasSuffix(lower, ".json.gz") {
-		return "application/json"
-	}
-	if strings.HasSuffix(lower, ".txt.gz") || strings.HasSuffix(lower, ".csv.gz") {
-		return "text/plain"
-	}
-	return "application/octet-stream"
-}
-
-func isNavigationAborted(err error) bool {
-	return err != nil && strings.Contains(err.Error(), "net::ERR_ABORTED")
 }
 
 // HandleTabDownload fetches a URL using the browser session for a tab identified by path ID.
